feat: integrate specta for type generation and improve API command handling

+++ src-tauri/src/activity.rs
@@ -1,11 +1,12 @@
 use chrono::{DateTime, Utc};
 use serde::Serialize;
+use specta::Type;
 use std::collections::VecDeque;
 
 pub const ACTIVITY_LOG_MAX: usize = 100;
 
 /// A single backend fetch event (prefetch, timer refresh, or manual cache refresh).
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Type)]
 pub struct FetchEvent {
     /// RFC3339 timestamp when the fetch completed.
     pub timestamp: String,
@@ -14,9 +15,9 @@ pub struct FetchEvent {
     /// Human-readable API endpoint summary, e.g. "/commodities_prices (per-entity, 120 calls)".
     pub endpoint: String,
     /// Number of rows/entries fetched.
-    pub row_count: usize,
+    pub row_count: u32,
     /// Wall-clock duration of the fetch in milliseconds.
-    pub duration_ms: u64,
+    pub duration_ms: u32,
     /// What triggered this fetch: "startup", "timer", or "manual".
     pub triggered_by: String,
     pub ok: bool,
@@ -24,7 +25,7 @@ pub struct FetchEvent {
 }
 
 /// The last price/entity lookup issued by the user (via a price command).
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Type)]
 pub struct LastUserAction {
     pub timestamp: String,
     /// Entity kind, e.g. "commodity", "vehicle".
@@ -34,10 +35,10 @@ pub struct LastUserAction {
     pub collection: String,
     /// "fresh", "stale", or "missing".
     pub source: String,
-    pub row_count: usize,
+    pub row_count: u32,
 }
 
-/// Central activity tracker stored on AppState.
+/// Central activity trackerstored on AppState.
 pub struct ActivityLog {
     pub events: VecDeque<FetchEvent>,
     pub last_bg_check_at: Option<DateTime<Utc>>,
